requirements.txt
flask
line-bot-sdk
requests
psycopg2-binary
